+++ stubs/classes.php
@@ -112,6 +112,7 @@ abstract class FloatX implements \JsonSerializable
 final class Float16 extends FloatX
 {
     //public function __toFloat(): float {}
+    public function toFloat32(): Float32 {}
 }
 
 /**
@@ -122,6 +123,7 @@ final class Float16 extends FloatX
 final class Float32 extends FloatX
 {
     //public function __toFloat(): float {}
+    public function toFloat16(): Float16 {}
 }
 
 /**